requirements.txt
flask
waitress
python-snappy
protobuf
PyMySQL
pymysql-pool==0.4.6
cryptography